# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Empty Spain API",
        "contact": {
            "name": "Grupo 5 Support Team",
            "email": "grupo5@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Admin access"
                ],
                "summary": "Gets all users records (except sensitive data such as password)",
                "operationId": "App\\Http\\Controllers\\DashboardController::getUsers",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token not provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/updateUser": {
            "post": {
                "tags": [
                    "Admin access"
                ],
                "summary": "Update user credentials",
                "operationId": "App\\Http\\Controllers\\DashboardController::updateUserCredentials",
                "requestBody": {
                    "description": "JSON containing user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "name",
                                    "surnames",
                                    "role",
                                    "is_disabled"
                                ],
                                "properties": {
                                    "email": {
                                        "$ref": "#/components/schemas/User/properties/email"
                                    },
                                    "name": {
                                        "$ref": "#/components/schemas/User/properties/name"
                                    },
                                    "surnames": {
                                        "$ref": "#/components/schemas/User/properties/surnames"
                                    },
                                    "role": {
                                        "$ref": "#/components/schemas/User/properties/role"
                                    },
                                    "is_disabled": {
                                        "$ref": "#/components/schemas/User/properties/is_disabled"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/deleteUser": {
            "post": {
                "tags": [
                    "Admin access"
                ],
                "summary": "Delete user permanently from the db",
                "operationId": "App\\Http\\Controllers\\DashboardController::deleteUser",
                "requestBody": {
                    "description": "JSON containing user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "$ref": "#/components/schemas/User/properties/id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/updateHighlighted": {
            "post": {
                "tags": [
                    "Admin access"
                ],
                "summary": "Updated municipio highlighted column",
                "operationId": "App\\Http\\Controllers\\DashboardController::updateHighlighted",
                "requestBody": {
                    "description": "JSON containing municipio id and highlighted value",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "highlighted"
                                ],
                                "properties": {
                                    "id": {
                                        "$ref": "#/components/schemas/Municipios/properties/id"
                                    },
                                    "highlighted": {
                                        "$ref": "#/components/schemas/Busqueda/properties/highlighted"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/municipiosWithHighlighted": {
            "get": {
                "tags": [
                    "Admin access"
                ],
                "summary": "Gets the highlighted municipios",
                "operationId": "App\\Http\\Controllers\\DashboardController::getMunicipiosWithHighlighted",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/mostSearchedMunicipios": {
            "get": {
                "tags": [
                    "Admin access"
                ],
                "summary": "Gets the top 10 most searched municipios",
                "operationId": "App\\Http\\Controllers\\DashboardController::getMostSearchedMunicipios",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/resetCcaaProvinciasMunicipios": {
            "get": {
                "tags": [
                    "Admin access"
                ],
                "summary": "Deletes municipios, provincias and CCAA columns from database and calls SqlSeeder to add them again",
                "operationId": "App\\Http\\Controllers\\DashboardController::ResetCcaaProvinciasMunicipios",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Reset success": {
                                            "type": "string",
                                            "example": "True"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/updateSearch": {
            "post": {
                "tags": [
                    "Admin access"
                ],
                "summary": "Updates fields of a search with provided data in the request body",
                "operationId": "App\\Http\\Controllers\\DashboardController::updateSearch",
                "requestBody": {
                    "description": "JSON containing municipio id and column to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "tripadvisor_info"
                                ],
                                "properties": {
                                    "id": {
                                        "$ref": "#/components/schemas/Municipios/properties/id"
                                    },
                                    "tripadvisor_info": {
                                        "$ref": "#/components/schemas/Busqueda/properties/tripadvisor_info"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/getDespoblacion": {
            "get": {
                "tags": [
                    "Admin access"
                ],
                "summary": "Get count of municpio_state of all the municipios",
                "operationId": "App\\Http\\Controllers\\DashboardController::getDespoblacion",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Reset success": {
                                            "type": "string",
                                            "example": "True"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
...
}